style: drop a redundant ulong cast on the invalid-handle sentinel
CK_INVALID_HANDLE is declared `public const ulong`, and the sessionId
parameter is already a ulong, so the cast converted ulong to ulong.
ObjectHandle.IsInvalid already compares without one; the other casts of
this constant target NativeCULong and are real conversions.

+++ src/KerckhoffsLabs.Security.Cryptography.Pkcs11/Internal/Pkcs11Session.cs
@@ -276,7 +276,7 @@ internal sealed class Pkcs11Session : IDisposable
 
         ArgumentNullException.ThrowIfNull(pkcs11Library);
 
-        if (sessionId == (ulong)CK.CK_INVALID_HANDLE)
+        if (sessionId == CK.CK_INVALID_HANDLE)
             throw new ArgumentException("Invalid handle specified", nameof(sessionId));
 
         _pkcs11Library = pkcs11Library;
